refactor loadAvailbleRoutes in menu service and add logic to allow users to switch b/w dropdown and custom link
