+++ src/app/features/dashboard-widgets/shortcut-dashboard-widget/shortcut-dashboard-settings.component/shortcut-dashboard-settings.component.html
@@ -11,6 +11,7 @@
           [item]="shortcut"
           (itemChange)="onShortcutItemChange($event, i)"
           [showLinkTooltip]="true"
+          [linkOptions]="availableRoutes"
         ></app-menu-item-form>
       </div>
     }
